queue messages on ProtoPhase::Init
the next thing that will probably need to be taken into account now is
using a different channel for client requests and replica requests, so
that the client requests do not starve the channel for the replica
network chatter.

we can use one channel per sub-protocol, and use a single receive()
behind them all, selecting fairly from each at random.